requirements.txt
# Bibliotecas principais
requests>=2.25.1
beautifulsoup4>=4.9.3
scikit-learn>=0.24.2
pandas>=1.3.0
numpy>=1.20.0
tldextract>=3.1.0
python-whois>=0.7.3
urllib3>=1.26.5

# Para a aplicação web
flask>=2.0.1
flask-wtf>=0.15.1

# Para testes
pytest>=6.2.5